Added types for src/Website/ClientRoute module
