--- src/Website/ClientRoute/ChainRoute/ChainList/chainConfiguration.js
@@ -1,10 +1,0 @@
-const ropstenNetwork = {
-  method: "wallet_switchEthereumChain",
-  params: [
-    {
-      chainId: "0x3",
-    },
-  ],
-};
-
-export { ropstenNetwork };
