--- src/Website/ClientRoute/ChainRoute/ChainList/chainList.js
@@ -1,70 +1,0 @@
-const chainList = {
-  0: "Kardia",
-  1: "Ethereum",
-  2: "Expanse",
-  3: "Ropsten Testnet",
-  8: "Ubiq",
-  10: "Optimism",
-  19: "Songbird",
-  20: "Elastos",
-  25: "Cronos",
-  30: "Rsk",
-  40: "Telos",
-  52: "Csc",
-  55: "Zyx",
-  56: "Binance",
-  57: "Syscoin",
-  60: "Gochain",
-  61: "Ethclassic",
-  66: "Okexchain",
-  70: "Hoo",
-  82: "Meter",
-  88: "Tomochain",
-  100: "Xdai",
-  106: "Velas",
-  108: "Thundercore",
-  122: "Fuse",
-  128: "Heco",
-  137: "Polygon",
-  200: "Xdaiarb",
-  246: "Energyweb",
-  250: "Fantom",
-  269: "Hpb",
-  288: "Boba",
-  321: "Kucoin",
-  336: "Shiden",
-  361: "Theta",
-  416: "Sx",
-  534: "Candle",
-  592: "Astar",
-  820: "Callisto",
-  888: "Wanchain",
-  1088: "Metis",
-  1284: "Moonbeam",
-  1285: "Moonriver",
-  2020: "Ronin",
-  2612: "Ezchain",
-  4181: "Phi",
-  4689: "Iotex",
-  5050: "Xlc",
-  5551: "Nahmii",
-  7777: "Nmactest",
-  8217: "Klaytn",
-  9001: "Evmos",
-  10000: "Smartbch",
-  103090: "Crystaleum",
-  32659: "Fusion",
-  42161: "Arbitrum",
-  42220: "Celo",
-  42262: "Oasis",
-  43114: "Avalanche",
-  71402: "Godwoken",
-  200625: "Akroma",
-  333999: "Polis",
-  1313161554: "Aurora",
-  1666600000: "Harmony",
-  0x11297108109: "Palm",
-  836542336838601: "Curio",
-};
-
-export { chainList };
